requirements.txt
numpy==2.3.2
scikit_fuzzy==0.5.0
paho-mqtt==2.1.0